Restful API 형식으로 매핑 uri 변경 및 crud 입력 밸리데이션 추가

+++ src/main/java/com/assignment/shoppingbasketserver/dao/ItemDao.java
@@ -10,8 +10,8 @@ import java.util.List;
 public interface ItemDao {
 
     int insertItem(ItemDto itemDto);
-    ItemDto selectItemByNo(Long itemNo);
+    ItemDto selectItemByNo(int itemNo);
     ItemDto selectItemByName(String itemName);
     List<ItemDto> selectItemLikeName(String itemName);
-    int deleteItem(Long itemNo);
+    int deleteItem(int itemNo);
 }
